Split TemplatedQueryForm.jsx into TemplatedListResultTable.jsx and TemplatedQueryForm.jsx
